refactor: extract agents last exam cli commands
